+++ RELEASE_NOTES.md
@@ -10,7 +10,7 @@
 - ODataAsyncResultFilter.DefaultConverter now returns correct counts
 - event publishing for non-event sourced entities (not transactionally safe and not saving to event store yet, though, so beware)
 - fixed BasicDomainModelConvention when using inheritance including abstract types
-- fixed binding of Hangfire configuration section
+- fixed binding of Hangfire and infrastructure configuration sections
 
 ## [1.4.0] - 2018-11-05